fix: move enums to services wakuext
