+++ tests-functional/clients/services/wakuext.py
@@ -88,6 +88,26 @@ class CommunityPermissionsAccess(Enum):
     MANUAL_ACCEPT = 3
 
 
+class CommunityTokenPermissionType:
+    BECOME_MEMBER = 1
+    BECOME_ADMIN = 2
+    BECOME_TOKEN_MASTER = 3
+    CAN_VIEW_CHANNEL = 4
+    CAN_VIEW_AND_POST_CHANNEL = 5
+
+
+class CommunityTokenType:
+    ERC20 = 1
+    ERC721 = 2
+    ENS = 3
+
+
+class CommunityTokenPrivilegesLevel:
+    OWNER_LEVEL = 1
+    MASTER_LEVEL = 2
+    COMMUNITY_LEVEL = 3
+
+
 class Error(Exception):
     def __init__(self, message):
         self.message = message
